parser/requirement.txt
Flask==2.3.2
PyMuPDF==1.22.3
python-dotenv==1.0.1
flask-cors==4.0.0
